Fix test typing
Previously there was a test which violated type-hints.
As we can not keep having the test _and_ type hints on that test, the
test itself was rewritten to test a semantically slightly different but
similar edge case.
The original test case is of course prevented when using type-hints (i.e
not passing `None` to a function that doesn't accept `None`)

Signed-off-by: Florian Wagner <florian.wagner@frequenz.com>