Use plain `float`s in the resampler buffer
When resampling we know we are working always with the same units, so
using a `Quantity` doesn't add any advantages to plain `float`s.
Instead, it makes the resampler slower, as it needs to do an extra
lookup each time a buffer value is needed and potentially avoids the
need to create `Quantity`s from the inputs.

The performance can be improved by about 25% (for resamples=1000
samples=1000, before 3.7s, now 2.8s) according to the resampling
benchmark.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>