Add a wall clock attached timer for the resampler
This timer uses the wall clock to trigger ticks and handles
discrepancies between the wall clock and monotonic time. Since sleeping
is performed using monotonic time, differences between the two clocks
can occur.

When the wall clock progresses slower than monotonic time, it is
referred to as *compression* (wall clock time appears in the past
relative to monotonic time). Conversely, when the wall clock progresses
faster, it is called *expansion* (wall clock time appears in the future
relative to monotonic time). If these differences exceed a configured
threshold, a warning is emitted.

If the difference becomes excessively large, it is treated as a *time
jump*. Time jumps can occur, for example, when the wall clock is
adjusted by NTP after being out of sync for an extended period. In such
cases, the timer resynchronizes with the wall clock and triggers an
immediate tick.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>