Add utility to mock the clocks for complex ticking tests
The `TimeDriver` class encapsulates the necessary mocks for
`datetime.datetime` and provides methods to manipulate wall clock time
during tests. This is particularly useful for testing components that
rely on both wall clock time (which can be adjusted by the system) and
monotonic time (which should always move forward).

We also register the util module so pytest rewrite asserts, as it uses
asserts to verify some stuff behaves as expected, and it is nicer to
get those asserts as test failures instead of programming errors.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>