Add custom comparison for `TickInfo` objects for pytest
Testing the ticking behavior is very complicated in itself, and when
there are failures, it is very difficult to understand what is
happening and where the differences between the expected and actual
tick times are.

To help with this, we add a custom comparison for `TickInfo` objects
that provides a detailed report of the differences between the
expected and actual tick times, including the sleep information for
each sleep.

This uses the special `pytest` hook `pytest_assertrepr_compare`:
https://docs.pytest.org/en/stable/reference/reference.html#pytest.hookspec.pytest_assertrepr_compare

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>