Add a new `ResamplerConfig2` that uses `WallClockTimerConfig`
This configuration object is exactly the same as `ResamplerConfig` but
replaces the `align_to` field with a new `timer_config` field that holds
a `WallClockTimerConfig` instance (or `None` to use the default). It
still inherits from `ResamplerConfig` to make the migration easier.

Because the `ResamplingFunction` also takes the config, a
`ResamplingFunction2` is also added to keep backwards compatibility with
any custom resampling function users might be using.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>