Test wall clock timer in resampled moving window tests
These tests are very tricky and there seems to be some inconsistency
between the monotonic and wall clock, the wall clock timer have less
samples in the buffer at the end. This is probably some border condition
because of the way we fake time.

We also had to tune how to do wall clock time shifting to make sure the
shift occurs before the timer wakes up.

Finally we adjusted sample sending so it always sends an odd number of
samples to make sure the resample is triggered at the points we expect
it, otherwise again we could end up in border condition that were not
the same for both timers.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>