Replace uses of `ComponentCategory` with the new types
The new API client provides a mapping of `ComponentCategory` and
`XxxType`s to Python classes. This commit updates most uses of component
categories with these classes.

This also includes some updates of imports for `Component` and
`ComponentCategory` and some changes on specific component type, like
`GridConnectionPoint`'s metadata.

Finally, the component graph interface to get components was also
updated to use types instead of categories, for which the API was
slightly changed to better represent the filtering criteria. The
keyword for arguments were changed from `component_ids` to
`matching_ids` and `component_categories` to `matching_types`.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>