Improve component graph's filtering
While we are changing the interface of component retrieval in the
component graph, we take the opportunity to make its interface more
generic.

We now accept any Iterable for `matching_ids` and `matching_types`, as
well as one individual items. Most uses of it filter by a single item
anyway so it is annoying to have to wrap it in a Iterable.

Most uses of those methods were updated to use the single-item overload.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>