Use new component states and state snapshots
The new state snapshots in the v0.18 client provide errors and warnings
using a different structure, so we need to account for that.

State code were also unified and there is no more specific enums for
cable state and other component-specific states, all states are part of
the `ComponentStateCode` enum and the component data can have many
state codes in it, while before it contained only one.

In previous versions EV chargers only used one `PLUGGED` or `LOCKED`
status but now the cable status was split in 2 connector ends: station
and EV, so now all combinations of those need to be accounted for. This
might change again on the API side as it ended up a bit messy.

Also the old component data wrappers now can't be frozen because the
`ComponentData` class is not frozen (this is to allow parsing from the
new telemetry messages easier for phases).

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>