Improve test resilience
We use `_recv_report_until()` when receiving more report to be more
tolerant to timing issues. For one side, if there is a race and some
report is sent with old values one last time, this keep receiving to
get the new (updated) report, and when there is an issue, it will use
the timeout in `_recv_report_until()` and not hang forever receiving
from the receiver.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>