Propagate Sample types through AST evaluation
This changes `AstNode.evaluate()` to return
`Sample[QuantityT] | QuantityT | None` instead of `float | None`.

This makes timestamps available to coalesce node, so it knows how to
synchronize newly started fallback streams with the primary streams.

This also requires a `create_method` to be passed to TelemetryStream,
for accurate typing, so the `Quantity` types from the resampler don't
get sent out as they are, in case of simple formulas, because there is
no top-level re-wrapping of the values in the formula evaluator
anymore.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>